spacewalk-java: Fix code for new Cobbler-API

+++ java/code/src/com/redhat/rhn/common/conf/ConfigDefaults.java
@@ -535,7 +535,7 @@ public class ConfigDefaults {
      * Returns the default virt disk size in GBs
      * @return the virt disk size
      */
-    public int getDefaultVirtDiskSize() {
+    public double getDefaultVirtDiskSize() {
         return Config.get().getInt(VIRT_DISK, 3);
     }
